# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apiv2.test/api",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/usertype": {
            "get": {
                "tags": [
                    "UserType"
                ],
                "summary": "Get list of all user types",
                "description": "Returns list of all user types",
                "operationId": "getUserTypeList",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Unsuccess"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UserType"
                ],
                "summary": "Creates a new user type",
                "description": "Creates a user type",
                "operationId": "storeUserType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Article's title",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success! User type created."
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/usertype/{id}": {
            "get": {
                "tags": [
                    "UserType"
                ],
                "summary": "Returns an user type giving its id",
                "description": "Returns an user type giving its id",
                "operationId": "showUserType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação com sucesso"
                    },
                    "400": {
                        "description": "Operação sem sucesso"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "UserType"
                ],
                "summary": "Updates a user type",
                "description": "Updates a user type",
                "operationId": "updateUserType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UserType id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "User type's title",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserType"
                ],
                "summary": "Deletes an user type giving its id",
                "description": "Deletes an user type giving its id",
                "operationId": "deleteUserType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação com sucesso"
                    },
                    "400": {
                        "description": "Operação sem sucesso"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "oauth2",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token/refresh",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}